360_igev_stereo/requirements.txt
glob2==0.7
hydra-core==1.3.2
matplotlib==3.9.2
numpy==2.1.1
omegaconf==2.3.0
opencv-python==4.10.0.84
pillow==10.4.0
timm==0.5.4
torch==2.4.1
torchvision==0.19.1
tqdm==4.66.5
wandb==0.18.0